pep8 repairs & class reorganisation
