+++ easybuild/tools/build_from_specsfile.py
@@ -1,6 +1,38 @@
 import yaml
-import os
-from easybuild.tools.specsfile.specsfile import Specsfile, SoftwareSpecs
+
+
+# general specs applicable to all commands
+class Specsfile(object):
+    def __init__(self):
+        self.easybuild_version = None
+        self.robot = False
+        self.software_list = []
+
+    # returns list of all commands - finished
+    def compose_eb_cmds(self):
+        eb_cmd_list = []
+        for sw in self.software_list:
+            eb_cmd_list.append(self.make_cmd(sw))
+        return eb_cmd_list
+
+    # single command
+    def make_cmd(self, sw):
+        eb_cmd = 'eb %s-%s-%s%s --robot=%s ' % (
+            sw.software, sw.version, sw.toolchain, sw.get_version_suffix(), self.robot)
+        return eb_cmd
+
+
+# single sw command
+class SoftwareSpecs(object):
+    def __init__(self, software, version, toolchain):
+        self.software = software
+        self.version = version
+        self.toolchain = toolchain
+        self.version_suffix = None
+
+    def get_version_suffix(self):
+        return self.version_suffix or ''
+
 
 # implement this to your own needs - to create custom yaml/json/xml parser
 class GenericSpecsParser(object):
@@ -13,17 +45,16 @@ class YamlSpecParser(GenericSpecsParser):
     @staticmethod
     def parse(filename):
         
-        # try:
-        with open(filename, 'r') as f:
-            spec_dict = yaml.safe_load(f)
+        try:
+            with open(filename, 'r') as f:
+                spec_dict = yaml.safe_load(f)
 
-        eb = Specsfile()
-        # except:
-        #     print("Cannot open file '" + filename + "'. Try to provide absolute path or adjust permissions.")
-        #     exit()
-        # loads all softwares' dictionaries from yaml
+            eb = Specsfile()
+        except FileNotFoundError:
+            print("Cannot open file '" + filename + "'. Try to provide absolute path or adjust permissions.")
+            exit()
         sw_dict = spec_dict["software"]
-        
+
         # assign software-specific EB attributes
         for software in sw_dict:
             try:
@@ -41,17 +72,24 @@ class YamlSpecParser(GenericSpecsParser):
                         eb.software_list.append(sw)
                         try:
                             version_info = sw_dict[software]['toolchains'][yaml_toolchain]['versions'][yaml_version]
-                            if version_info['versionsuffix'] != None:
-                                sw.version_suffix=version_info['versionsuffix']
-                        except:
+                            if version_info['versionsuffix'] is not None:
+                                sw.version_suffix = version_info['versionsuffix']
+                        except (KeyError, TypeError, IndexError) as e:
                             continue
-                        
-            except:
+            except (KeyError, TypeError, IndexError) as ex:
                 print('Software ' + str(software) + ' has wrong yaml structure!')
 
         # assign general EB attributes
         eb.easybuild_version = spec_dict.get('easybuild_version', None)
         eb.robot = spec_dict.get('robot', False)
-        
         return eb
 
+
+def handle_specsfile(filename):
+    
+    eb = YamlSpecParser.parse(filename)
+    
+    eb_cmds = eb.compose_eb_cmds()
+    
+    for x in eb_cmds:
+        print(x)
